refactor: creates new `makeCommentAfterFixes` function
